drm/vmwgfx: Remove initialization of connector status
Remove the connector-status initialization from several of vmwgfx's
functions. It is not required by the driver or DRM helpers.

DRM initializes the connector to unknown status in __drm_connector_init()
and reads the physical status when DRM clients or the user needs the
information, or if it detects/polls a change to the status.

Signed-off-by: Thomas Zimmermann <tzimmermann@suse.de>
Reviewed-by: Zack Rusin <zack.rusin@broadcom.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20241105132045.471856-1-tzimmermann@suse.de